Exact match srch priority >> r-t-c-l.sh
This splits a for loop into something like "sibling loops" because exact matches are easy enough to check for, and common enough when it comes to CODEOWNERS. If exact matches don't work, then and only then should we do more advanced search. This is how it functioned somewhat already but not optimally.